+++ azure_monitor/src/azure_monitor/options.py
@@ -27,6 +27,7 @@ class ExporterOptions(BaseObject):
     Args:
         connection_string: Azure Connection String.
         instrumentation_key: Azure Instrumentation Key.
+        proxies: Proxies to pass Azure Monitor request through.
         storage_maintenance_period: Local storage maintenance interval in seconds.
         storage_max_size: Local storage maximum size in bytes.
         storage_path: Local storage file path.
@@ -36,8 +37,8 @@ class ExporterOptions(BaseObject):
 
     __slots__ = (
         "connection_string",
-        "endpoint",
         "instrumentation_key",
+        "proxies",
         "storage_maintenance_period",
         "storage_max_size",
         "storage_path",
@@ -49,6 +50,7 @@ class ExporterOptions(BaseObject):
         self,
         connection_string: str = None,
         instrumentation_key: str = None,
+        proxies: typing.Dict[str, str] = None,
         storage_maintenance_period: int = 60,
         storage_max_size: int = 50 * 1024 * 1024,
         storage_path: str = None,
@@ -64,6 +66,7 @@ class ExporterOptions(BaseObject):
             )
         self.connection_string = connection_string
         self.instrumentation_key = instrumentation_key
+        self.proxies = proxies
         self.storage_maintenance_period = storage_maintenance_period
         self.storage_max_size = storage_max_size
         self.storage_path = storage_path
@@ -74,6 +77,7 @@ class ExporterOptions(BaseObject):
         self._validate_instrumentation_key()
 
     def _initialize(self) -> None:
+        # connection string and ikey
         code_cs = parse_connection_string(self.connection_string)
         code_ikey = self.instrumentation_key
         env_cs = parse_connection_string(
@@ -102,6 +106,11 @@ class ExporterOptions(BaseObject):
             or "https://dc.services.visualstudio.com"
         )
         self.endpoint = endpoint + "/v2/track"
+
+        # proxies
+        if self.proxies is None:
+            self.proxies = '{}'
+
 
     def _validate_instrumentation_key(self) -> None:
         """Validates the instrumentation key used for Azure Monitor.