+++ azure_monitor/tests/test_options.py
@@ -241,6 +241,24 @@ class TestOptions(unittest.TestCase):
             options.endpoint, "https://dc.services.visualstudio.com/v2/track"
         )
 
+    def test_process_options_proxies_default(self):
+        options = ExporterOptions()
+        options.proxies = "{}"
+        common.process_options(options)
+
+        self.assertEqual(options.proxies, "{}")
+
+    def test_process_options_proxies_set_proxies(self):
+        options = ExporterOptions()
+        options.connection_string = None
+        options.proxies = '{"https": "https://test-proxy.com"}'
+        common.process_options(options)
+
+        self.assertEqual(
+            options.proxies,
+            '{"https": "https://test-proxy.com"}'
+        )
+
     def test_parse_connection_string_invalid(self):
         self.assertRaises(
             ValueError, lambda: ExporterOptions(connection_string="asd")