+++ azure_monitor/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Implement proxies in exporter configuration
+- ([#92](https://github.com/microsoft/opentelemetry-azure-monitor-python/pull/92))
 - Remove dependency metrics from auto-collection
 - ([#92](https://github.com/microsoft/opentelemetry-azure-monitor-python/pull/92))